Fixed initial conversation loading bug
