+++ src/App.vue
@@ -14,24 +14,10 @@
 </template>
 <script>
 import navbar from "../components/NavBar";
-import api from "../api";
 
 export default {
   components: {
     navbar
-  },
-  async mounted() {
-    try {
-      const {
-        data: lastConversation
-      } = await api.conversation.getLastConversation();
-      this.$store.dispatch(
-        "conversation/setActiveConversation",
-        lastConversation
-      );
-    } catch (error) {
-      console.error(error);
-    }
   },
   data() {
     return {
